# backend/docs/swagger.yaml
definitions:
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.AvailableApparatNumbersResponse:
    properties:
      available:
        items:
          type: integer
        type: array
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BacnetObjectResponse:
    properties:
      alarm_definition_id:
        type: string
      alarm_type_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      field_device_id:
        type: string
      gms_visible:
        type: boolean
      hardware_quantity:
        type: integer
      hardware_type:
        type: string
      id:
        type: string
      notification_class_id:
        type: string
      optional:
        type: boolean
      software_number:
        type: integer
      software_reference_id:
        type: string
      software_type:
        type: string
      state_text_id:
        type: string
      text_fix:
        type: string
      text_individual:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse:
    properties:
      error:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.FieldDeviceListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.FieldDeviceResponse'
        type: array
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.FieldDeviceResponse:
    properties:
      apparat:
        $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ApparatResponse'
      apparat_id:
        type: string
      apparat_nr:
        type: integer
      bacnet_objects:
        items:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BacnetObjectResponse'
        type: array
      bmk:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      specification:
        $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.SpecificationResponse'
      specification_id:
        type: string
      sps_controller_system_type:
        allOf:
        - $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.SPSControllerSystemTypeResponse'
        description: Embedded related entities for display
      sps_controller_system_type_id:
        type: string
      system_part:
        $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.SystemPartResponse'
      system_part_id:
        type: string
      text_fix:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.UpdateFieldDeviceRequest:
    properties:
      apparat_id:
        type: string
      apparat_nr:
        maximum: 99
        minimum: 1
        type: integer
      bacnet_objects:
        items:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BacnetObjectInput'
        type: array
      bmk:
        maxLength: 10
        type: string
      description:
        maxLength: 250
        type: string
      object_data_id:
        type: string
      sps_controller_system_type_id:
        type: string
      system_part_id:
        type: string
      text_fix:
        maxLength: 250
        type: string
    required:
    - system_part_id
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/facility/field-devices:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 300
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      - description: Order by (created_at,sps_system_type,bmk,description,apparat_nr,apparat,system_part,spec_supplier,spec_brand,spec_type,spec_motor_valve,spec_size,spec_install_loc,spec_ph,spec_acdc,spec_amperage,spec_power,spec_rotation)
        in: query
        name: order_by
        type: string
      - description: Order direction (asc, desc)
        in: query
        name: order
        type: string
      - description: Filter by building ID
        in: query
        name: building_id
        type: string
      - description: Filter by control cabinet ID
        in: query
        name: control_cabinet_id
        type: string
      - description: Filter by SPS controller ID
        in: query
        name: sps_controller_id
        type: string
      - description: Filter by SPS controller system type ID
        in: query
        name: sps_controller_system_type_id
        type: string
      - description: Filter by project ID
        in: query
        name: project_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.FieldDeviceListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: List field devices with pagination and filtering
      tags:
      - facility-field-devices
  /api/v1/facility/field-devices/{id}:
    delete:
      parameters:
      - description: Field Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Delete a field device
      tags:
      - facility-field-devices
    get:
      parameters:
      - description: Field Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.FieldDeviceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Get a field device by ID
      tags:
      - facility-field-devices
    put:
      consumes:
      - application/json
      parameters:
      - description: Field Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Field Device data
        in: body
        name: field_device
        required: true
        schema:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.UpdateFieldDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.FieldDeviceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Update a field device
      tags:
      - facility-field-devices
  /api/v1/facility/field-devices/{id}/bacnet-objects:
    get:
      parameters:
      - description: Field Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BacnetObjectResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: List bacnet objects for a field device (hydration)
      tags:
      - facility-field-devices
  /api/v1/facility/field-devices/{id}/specification:
    post:
      consumes:
      - application/json
      parameters:
      - description: Field Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Specification data
        in: body
        name: specification
        required: true
        schema:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.CreateFieldDeviceSpecificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.SpecificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Create specification for a field device
      tags:
      - facility-field-devices
    put:
      consumes:
      - application/json
      parameters:
      - description: Field Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Specification data
        in: body
        name: specification
        required: true
        schema:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.UpdateFieldDeviceSpecificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.SpecificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Update specification for a field device
      tags:
      - facility-field-devices
  /api/v1/facility/field-devices/available-apparat-nr:
    get:
      parameters:
      - description: SPS Controller System Type ID
        in: query
        name: sps_controller_system_type_id
        required: true
        type: string
      - description: Apparat ID
        in: query
        name: apparat_id
        required: true
        type: string
      - description: System Part ID
        in: query
        name: system_part_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.AvailableApparatNumbersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: List available apparat numbers for field devices
      tags:
      - facility-field-devices
  /api/v1/facility/field-devices/bulk-delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bulk delete request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BulkDeleteFieldDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BulkDeleteFieldDeviceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Bulk delete multiple field devices
      tags:
      - facility-field-devices
  /api/v1/facility/field-devices/bulk-update:
    patch:
      consumes:
      - application/json
      description: Updates multiple field devices in a single operation. Supports
        nested specification and BACnet objects updates.
      parameters:
      - description: Bulk update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BulkUpdateFieldDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BulkUpdateFieldDeviceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Bulk update multiple field devices
      tags:
      - facility-field-devices
  /api/v1/facility/field-devices/multi-create:
    post:
      consumes:
      - application/json
      description: Creates multiple field devices with independent validation. Returns
        detailed results for each device. To link created devices to a project, use
        the CreateProjectFieldDevice endpoint with the returned field device IDs.
      parameters:
      - description: Multi-create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.MultiCreateFieldDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.MultiCreateFieldDeviceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Create multiple field devices in a single operation
      tags:
      - facility-field-devices
  /api/v1/facility/field-devices/options:
    get:
      description: Returns all apparats, system parts, object datas and their relationships
        in a single call. This returns global templates (object data where project_id
        is null and is_active = true).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.FieldDeviceOptionsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Get all metadata needed for creating/editing field devices
      tags:
      - facility-field-devices
...
swagger: "2.0"